refactor component soure file to index.js for short import
