+++ albums/src/components/AlbumList/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { View, StyleSheet } from 'react-native';
 import axios from 'axios';
-import AlbumDetail from '../AlbumDetail/AlbumDetail';
+import AlbumDetail from '../AlbumDetail';
 
 class AlbumList extends Component {
   state = { albums: [] };
