+++ albums/src/components/Content/index.js
